Moved I2C file to common
